Do not set-iam-policy in case there is no change in audit configs

+++ common/modules/gcp-project/audit.tf
@@ -11,19 +11,25 @@ resource "null_resource" "add_audit_config" {
 
   provisioner "local-exec" {
     command = <<EOF
-      echo "Applying audit config..."
 
-      auditConfigs=$(cat ${path.module}/resources/auditConfigs.json | jq -c -r .)
-      bindings=$(gcloud projects get-iam-policy ${google_project.project.project_id} --format json | jq -c -r .bindings)
+      auditConfigs=$(cat ${path.module}/resources/auditConfigs.json | jq -c -r -S .)
+      policy=$(gcloud projects get-iam-policy ${google_project.project.project_id} --format json)
+      policy_bindings=$(echo $policy | jq -c -r .bindings)
+      policy_auditConfigs=$(echo $policy | jq -c -r -S .auditConfigs)
 
-      jq -n -r \
-        --argjson auditConfigs "$auditConfigs" \
-        --argjson bindings "$bindings" \
-        '{"auditConfigs":$auditConfigs,"bindings":$bindings}' > ${path.module}/${google_project.project.project_id}.iam.policy.json
+      if [ "$(echo $auditConfigs | md5sum)" != "$(echo $policy_auditConfigs | md5sum)" ]; then
+        jq -n -r \
+          --argjson auditConfigs "$auditConfigs" \
+          --argjson bindings "$policy_bindings" \
+          '{"auditConfigs":$auditConfigs,"bindings":$bindings}' > ${path.module}/${google_project.project.project_id}.iam.policy.json
 
-      yes | gcloud projects set-iam-policy ${google_project.project.project_id} ${path.module}/${google_project.project.project_id}.iam.policy.json
+        echo "Applying audit configs..."
+        gcloud -q projects set-iam-policy ${google_project.project.project_id} ${path.module}/${google_project.project.project_id}.iam.policy.json
 
-      rm ${path.module}/${google_project.project.project_id}.iam.policy.json
+        rm ${path.module}/${google_project.project.project_id}.iam.policy.json
+      else
+        echo "Audit configs are up-to-date..."
+      fi
     EOF
   }
 }
